requirements.txt
torch
torchvision
opencv-python
ultralytics
python-dotenv
roboflow
streamlit
flask